New CtranIb UT: AbortWaitNotify
Summary:
This diff introduces a new unit test for the `CtranIb` class, specifically testing the behavior of the `waitNotify` operation when an abort is triggered.

The test sets up two ranks and waits for them to exchange information using the `waitNotify` operation. However, the second rank never sends the notification, causing the first rank to wait indefinitely. The test then aborts the `waitNotify` and verifies that the VC is not created.

This test ensures that `CtranIb` process can handle unexpected behavior during the `waitNotify` operation and gracefully abort the operation.

Reviewed By: arttianezhu

Differential Revision: D87670526

fbshipit-source-id: 449f384e05781ce2f6ac395d379a11584c6899fb