Fix timeoutWatchdog failing cuda graph collectives
Summary: Previously, if we run a cuda graph collective after a non-cuda graph collective, there is a chance that it will fail due to timeoutWatchdog calling `cudaEventQuery` during the graph capture, which is not allowed by default and will cause all the subsequent cuda calls on this graph to fail. Switch the timeoutThread to cudaStreamCaptureModeThreadLocal mode to solve this problem.

Reviewed By: tanquer

Differential Revision: D89092598

fbshipit-source-id: 90793753283a4947943797631f018e1980a9c85f