Ensuring ctran support functions return false for algo=orig case
Summary:
Previously had not made sure that the ctranBroadcastSupport and ctranReduceScatterSupport functions returned false when the NCCL_BROADCAST_ALGO::orig and NCCL_REDUCESCATTER_ALGO::orig values were passed in (these are invalid queries since they are not ctran algos). Making sure of that in this diff.
Now that the ctranBroadcastSupport function requires the NCCL_BROADCAST_ALGO to be set to a ctran algo, we also make sure that NCCL_BROADCAST_ALGO is not set to orig within MCCL.

Reviewed By: Regina8023

Differential Revision: D88814199

fbshipit-source-id: c51cda76beebd54d5ecd7449ec1c377e7a439b89