tcp-devmem: move to pointers for unpack pool
Summary: Sheldon mentioned in D89059982 that with an internal vector we might hit thread safety issues unless we lock the vector everywhere (which we don't want to). Replace pool indices with pool pointers (which are "owned" by the callers). I was using indices initially because it was easier to debug, but having an internal vector to map indices to the pointers is indeed excessive.

Reviewed By: sheldnwu

Differential Revision: D89396751

fbshipit-source-id: 18168c1ee6e6bb13e8bd07bdad69c8b5b013c02b