Adding algo to ctran AllReduce
Summary:
Part of larger effort to add algo as a parameter for all collective APIs, allowing for future overwriting of algo as necessary and also allowing for consistency across collectives.
Prior to this diff, ctran AllReduce did not take algo as a required parameter in its API or in its support check function. This diff adds algo as a parameter to those functions and updates all the places that call those functions accordingly.
Currently, there is a hang for ctarg and ctring algos, so we return false from the support function for these algos. In the future if a new algo is added as an option for AllReduce, the API and support check function should be updated as well.

Reviewed By: arttianezhu

Differential Revision: D88813973

fbshipit-source-id: dbbd3ba6d7878b09922eb1f596ccb9d799c05a37